requirements.txt
openai-whisper==20231117
google-generativeai==0.8.3
python-dotenv==1.0.1
streamlit==1.40.1
torch>=1.9.0